Fixed order of insertion for elements onto elemStacks with preexisting elements of same level.
Discovered this bug when testing the editor. Placing a Placeable, moving off it, then moving back on
would cause the editor to slip under the placed object on the stack. addElem() now iterates from the right
and compares height by >= instead of > to prioritize rendering of most recently added elements.