Beginning of SDL2 ground-up refactor
In the course of doing development with curses, I realized that it was made for TUI development
first and game development pretty much not-at-all. To build a fully fleshed out game, it would take a lot
of (arguably) unnecessary labor and would leave me with a non-cross-platform product. Switching to SDL2
allows for much greater flexibiltiy and direct support for game development, texture rendering, tile-maps, etc.
while still allowing me to practice good memory management skills and build out a class hierarchy

In this commit:
* Basic game class, rendering of a simple window
* Class hierarchy skeleton for the various elements of the game
* Basic event handling (only handles ESC key rn)
* Ways to wrap and handle texture rendering
* .gitignore update